[issue-2223] - Lineage Callback for any operator (#2225)
* Fix logging

* Reorganise lineage

* Add callback

* Add metadata config

* Fix type smell
