+++ ingestion/src/metadata/ingestion/ometa/mixins/lineage_mixin.py
@@ -111,6 +111,6 @@ class OMetaLineageMixin(Generic[T]):
         except APIError as err:
             logger.error(
                 f"Error {err.status_code} trying to GET linage for "
-                + f"{entity.__class__.__name__} and {path}"
+                + f"{entity.__name__} and {path}"
             )
             return None
